refact: move stubs and remove unused depencies
