+++ tests/Integration/EagerLoadingTest.php
@@ -1,13 +1,11 @@
 <?php
 namespace Mongolid\Tests\Integration;
 
-use MongoDB\BSON\ObjectId;
 use Mongolid\Container\Container;
-use Mongolid\Tests\Stubs\Price;
 use Mongolid\Tests\Stubs\Legacy\Product;
-use Mongolid\Tests\Stubs\ReferencedUser;
-use Mongolid\Tests\Stubs\Shop;
-use Mongolid\Tests\Stubs\Sku;
+use Mongolid\Tests\Stubs\Legacy\Shop;
+use Mongolid\Tests\Stubs\Legacy\Sku;
+use Mongolid\Tests\Stubs\Price;
 use Mongolid\Util\CacheComponent;
 use Mongolid\Util\CacheComponentInterface;
