+++ tests/Stubs/Legacy/LegacyRecordStudent.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Mongolid\Tests\Stubs;
+namespace Mongolid\Tests\Stubs\Legacy;
 
 use Mongolid\LegacyRecord;
