+++ tests/Stubs/Legacy/LegacyRecordUser.php
@@ -1,5 +1,5 @@
 <?php
-namespace Mongolid\Tests\Stubs;
+namespace Mongolid\Tests\Stubs\Legacy;
 
 use Mongolid\LegacyRecord;
