+++ tests/Stubs/Price.php
@@ -2,7 +2,6 @@
 namespace Mongolid\Tests\Stubs;
 
 use Mongolid\Model\AbstractModel;
-use Mongolid\Model\Relations\ReferencesOne;
 
 class Price extends AbstractModel
 {
