+++ tests/Unit/DynamicSchemaTest.php
@@ -5,7 +5,6 @@ namespace Mongolid;
 use Mockery as m;
 use Mongolid\Schema\DynamicSchema;
 use Mongolid\Schema\Schema;
-use Mongolid\TestCase;
 
 class DynamicSchemaTest extends TestCase
 {
