+++ tests/Unit/Query/EagerLoader/ExtractorTest.php
@@ -1,18 +1,11 @@
 <?php
 namespace Mongolid\Query\EagerLoader;
 
-use ArrayIterator;
-use MongoDB\BSON\ObjectId;
-use Mongolid\Container\Container;
-use Mongolid\Cursor\SchemaCursor;
-use Mongolid\Cursor\SchemaEmbeddedCursor;
 use Mongolid\Query\EagerLoader\Exception\EagerLoaderException;
 use Mongolid\TestCase;
-use Mockery as m;
-use Mongolid\Tests\Stubs\Price;
 use Mongolid\Tests\Stubs\Legacy\Product;
-use Mongolid\Tests\Stubs\Shop;
-use Mongolid\Util\CacheComponentInterface;
+use Mongolid\Tests\Stubs\Legacy\Shop;
+use Mongolid\Tests\Stubs\Price;
 
 class ExtractorTest extends TestCase
 {
@@ -48,7 +41,7 @@ class ExtractorTest extends TestCase
             ],
             'shop' => [
                 'key' => 'skus.shop_id',
-                'model' => 'Mongolid\Tests\Stubs\Shop',
+                'model' => 'Mongolid\Tests\Stubs\Legacy\Shop',
                 'ids' => [
                     12345 => 12345,
                 ],
